Half Year Part Time Development Commit.
As mentioned in my previous commit, I have put aside github related activities and have been working on the project in any of my spare time - trying to develop it piece by piece. The strategy was is to interate quickly, as a solo developer the importance of git did not seem as essential as the ability to make changes and not ask or answer question. This allowed me to get maximal coding done in the limited time available. On average working 30~ hours a week in a physical labour job and commiting 20~ hours towards this project.

This is a titanic commit which mainly compromised of reoganizing ALL the original 'caoco' compiler module into a new 'cnd' compiler module. This was a significant chunk of effort as the project is already huge. The main purpose of this refactoring was due to the way project was structured. The headers were themed and mostly self contained allowing for great locality of reference- the downside was when the project got to a certain size: this structure ended up rigid and hard to read, the project is still at an early stage and the goal is to be accepting of major sweeping changes. Splitting up and re-organizing headers while maintaining the functionality took approximatley 4 months.

The next chunk of effort went towards studying automata theory, context free grammars and lr parsing techniques for 2 months. I felt this knowledge was what i lacked in fully completing the compiler's frontend with confidence. I had gotten stuck, unable to find a way to elegantly hande the parsing of complex primary expressions. It seems all my attempts were naive hacks and inefficient - would break apart once an edge case was encountered. Im speaking specifically on the caoco::parser::LRParser class which was actually a generic 'shift-reduce' parser.  Armed with my research i revisited and for a 3rd time went over every single detail of the entire parse implementation, cleaning and improvig the implementation. I have created a partial CFG and EBNF grammar for the C& language. Using the grammar as reference the lexer and parser I managed to use 'recursive descent' techniques to implement the same but must more code-healthy primary expression parser. Furthermore ,as a bonus, lexer and parser are now fully consteval evaluable.

Lastly i spent 2 months revieqing previous CMake knowledge. I applied my studies to improve and upgrade the 'minitest' module. Providing integration with CTest and MSVC test explorer. Significantly improved the minitest cmake script, following cmake guidelines and up to date practices. While already working on the module I also decided to add C++14 support now so that future migration will be easier(it is always the goal to support the lowest C++ possible). Ovaerall improved the  implementation of the minitest framework, mostly removing bugs and confusing code.

Currently as of this commint I am working on integrating the new minitest upgrade into this codebase. In addition to adding cmake support I also improved the way the library itself is implemented. Unfortunatley these changes introduce major api breaking changes which means ALL of the unit tests I ever wrote using the previous minitest library version will have to be refactored.

The next step is to clean up the top level project CMakeLists.txt, its filled with garbage/dummy tests from my experiments with adding CTest to minitest.

While CMake is fresh on my mind I think it would be optimal to work on this section of the project. It will give me time to look back and reflect on the goals of this programming language - as I have lost sight of the initial inspiration. Its essential to find a direction - or I will never feel a sense of completion on this project.